[11.0][MIG] dbfilter_from_header
updates were made in #1137 by @TimLai125

and small lint fixes by @yelizariev
