+++ dbfilter_from_header/__manifest__.py
@@ -1,12 +1,12 @@
-# -*- coding: utf-8 -*-
 # © 2013  Therp BV
 # © 2014  ACSONE SA/NV
+# Copyright 2018 Quartile Limited
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 {
     "name": "dbfilter_from_header",
     "summary": "Filter databases with HTTP headers",
-    "version": "10.0.1.0.0",
+    "version": "11.0.1.0.0",
     "author": "Therp BV,Odoo Community Association (OCA)",
     "license": "AGPL-3",
     "complexity": "normal",
